move components to /side-effect, /pure, /page
