+++ src/components/pure/Card.tsx
