+++ src/components/pure/Sidebar/Box/FounderItem.tsx
