+++ src/pages/docs.tsx
@@ -1,4 +1,4 @@
-import Page from '@/containers/Page'
+import Page from '@/components/side-effects/Page'
 
 const Docs = () => <Page>Documentation</Page>
 export default Docs
